Finally found a way to have a button in a wysiwyg editor (Iframes)
Bascially, the problem I was having was that whenever I clicked on a button such as bold, it would undo my selection. Most solutions to this involve Jquery, which conflicted with react. The workaround is to use an Iframe, but because I was using codepen and jsfiddle to test initially, I was getting weird glitches due to those websites putting the html code in an iframe. I didn't realize this because Iframe starting tags are hidden in chrome inspect element and replaced with `#document`, which makes it look like you can add a <html> tag in the middle of the document, which actually does nothing.