~Changing Goals~
After a bunch of testing, it seems like it would make much more sense to use the older storage code because the new code doesn't work as reliably and is taking longer.

After winter break, the main order of goals will be:
1: move all the old CSS to .SCSS,  and then possibly move .SCSS to .SASS
2: "clean up" the old storage code. don't use the new IPC thing, but instead use the specific strings.
3: Get saving to work and get it ready to be used in rounds
4: refactor storage after release

Will wait until after exams to do this.